+++ Source/Public/ArhqenCognitionEngine/Core/Application.h
@@ -2,6 +2,7 @@
 
 #include "ArhqenCognitionEngine/Core/AppConfig.h"
 #include "ArhqenCognitionEngine/Core/Assets/AceAssetRegistry.h"
+#include "ArhqenCognitionEngine/Core/Assets/AceAssetDirectoryWatcher.h"
 #include "ArhqenCognitionEngine/Core/ExitCode.h"
 #include "ArhqenCognitionEngine/Core/Logger.h"
 #include "ArhqenCognitionEngine/Core/RuntimeClock.h"
@@ -39,6 +40,7 @@ namespace am::core
         RuntimeClock clock_;
         TaskQueue tasks_;
         am::core::assets::AssetRegistry assetRegistry_;
+        am::core::assets::AssetDirectoryWatcher assetDirectoryWatcher_;
         am::renderer::NativeWindow window_;
         am::renderer::Dx12Renderer renderer_;
         am::renderer::UiDrawList uiDrawList_;
@@ -51,5 +53,9 @@ namespace am::core
         bool loggerReady_ = false;
         bool exitRequested_ = false;
         std::string exitReason_;
+        bool assetChangesPending_ = false;
+        bool assetFullRescanPending_ = false;
+        double assetChangeQuietSeconds_ = 0.0;
+        double assetChangePendingAgeSeconds_ = 0.0;
     };
 }